LayoutTests/editing/pasteboard/paste-text-012-expected.txt

See this bug: <rdar://problem/3918712> "Paste as Quotation" in Mail just pastes (<blockquote> tag seems to be lost).

To manually test, copy everything in the first editable box and paste it in the second box. You should see "foo" and a new line inside a red box.

foo
foo

execCopyCommand: <div id="test" class="editing"><div><blockquote>foo</blockquote></div></div> <div class="editing"></div>
execPasteCommand: <div id="test" class="editing"><div><blockquote>foo</blockquote></div></div> <div class="editing"><div id="test" class="editing"><blockquote>foo</blockquote><div><br></div></div></div>
